+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This changelog covers what's changed in the MPDX APIs.
 
+## 08 June 2017
+- Tnt import: improve setting task completed status, and timezone parsing
+
 ## 06 June 2017
 - Added wildcard search filter on '/api/v2/contacts/people' endpoint.
 - Added location to list of task attributes displayed on '/api/v2/tasks' endpoint.